services/intelligence/web_intelligence_service/pyproject.toml
[tool.poetry]
name = "web-intelligence-service"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
firecrawl-py = "^1.5.0"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
